better error handling with websockets
I thought that I could return an HTTP error before the websocket
upgrade which would be visible from the client. This, however,
does not seem to be the case. Instead, the client apparently gets the
same error no matter if we got an HTTP error, or a connection to the
server cannot be established.

- Rewrite the backend so that when a player registers it returns an
error by closing the websocket.
- Use that to inform the user whether the game does not exist or there
was an error contacting the server.